+++ crates/flui-widgets/src/text/mod.rs
@@ -11,7 +11,7 @@ pub mod text_field;
 
 pub use controller::TextEditingController;
 pub use default_text_style::DefaultTextStyle;
-pub use editable_text::{EditableText, EditableTextState};
+pub use editable_text::{EditableText, EditableTextState, SubmitCallback};
 pub use rich_text::RichText;
 pub use text::Text;
-pub use text_field::{TextField, TextFieldState};
+pub use text_field::{RawTextField, RawTextFieldState};